+++ CHANGELOG.md
@@ -1,0 +1,3 @@
+2.0.0
+
+- The `/u` endpoint has been renamed to `/fromTwitter`
